updated find and pull requests for delta set bugfixes
